pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tooluse"
version = "0.1.0"
description = "Offline synthetic multi-agent tool-use conversation generator"
requires-python = ">=3.10"
dependencies = [
    "networkx>=3.0",
    "mem0ai",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-timeout>=2.0",
]

[project.scripts]
tooluse = "tooluse.cli:main"

[tool.setuptools.packages.find]
where = ["ConvTool"]

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/e2e"]
